fix: same track not replaying if paused + import-export hooks
