+++ module/pf2/index.js
@@ -1,2 +1,3 @@
 import './settings.js';
 import './trait-manager.js';
+import './transfer.js';
